+++ modules/app-android/src/androidMain/kotlin/org/archivekeep/app/android/MainActivity.kt
@@ -3,6 +3,8 @@ package org.archivekeep.app.android
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
+import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
 import androidx.compose.runtime.remember
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -13,6 +15,7 @@ import org.archivekeep.app.desktop.domain.wiring.ApplicationProviders
 import org.archivekeep.app.desktop.ui.MainWindowContent
 
 class MainActivity : AppCompatActivity() {
+    @OptIn(ExperimentalMaterial3WindowSizeClassApi::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -25,6 +28,7 @@ class MainActivity : AppCompatActivity() {
             ApplicationProviders(environment) {
                 MainWindowContent(
                     isFloating = false,
+                    windowSizeClass = calculateWindowSizeClass(this),
                     onCloseRequest = { this@MainActivity.onBackPressed() },
                 )
             }